pyproject.toml
[tool.poetry]
name = "rubberfroggy"
version = "0.1.0"
description = ""
authors = ["Will Humphlett <will@humphlett.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
pystray = "^0.19.3"
screeninfo = "^0.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "rubberfroggy:main"

[tool.black]
line-length = 120
